src/main/java/com/mpalourdio/springboottemplate/service/PasswordService.java
/*
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
 * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
 * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */

package com.mpalourdio.springboottemplate.service;

import org.springframework.security.crypto.bcrypt.BCrypt;
import org.springframework.security.crypto.codec.Base64;

public class PasswordService {

    public String generateEncryptedPasword(final String plainTextPassword) {
        return BCrypt.hashpw(plainTextPassword, BCrypt.gensalt());
    }

    public Boolean isPasswordValid(final String plainTextPassword, final String hash) {
        return BCrypt.checkpw(plainTextPassword, hash);
    }

    public String encodeStringToBase64(final String toEncode) {
        return new String(Base64.encode(toEncode.getBytes()));
    }

    public String decodeBase64String(final String base64EncodedString) {
        return new String(Base64.decode(base64EncodedString.getBytes()));
    }
}
